* renamed cli.logging to cli.cli_logging to prevent a legacy python naming conflict * linted imports and killed some dead ones
