+++ great_expectations/cli/project.py
@@ -2,8 +2,12 @@ import sys
 
 import click
 
-from great_expectations import exceptions as ge_exceptions, DataContext
-from great_expectations.cli.util import cli_message, _offer_to_install_new_template
+from great_expectations import DataContext
+from great_expectations import exceptions as ge_exceptions
+from great_expectations.cli.util import (
+    _offer_to_install_new_template,
+    cli_message,
+)
 
 
 @click.group()
